requirements.txt
torch
torchvision
torchsummary
tqdm
pandas
torchmetrics
pytorch-ignite
tensorboard